requirements.txt
altgraph==0.17.3
certifi==2023.5.7
charset-normalizer==3.1.0
coloredlogs==15.0.1
flatbuffers==23.5.26
humanfriendly==10.0
idna==3.4
imageio==2.31.1
lazy_loader==0.3
llvmlite==0.40.1
mpmath==1.3.0
networkx==3.1
numba==0.57.1
numpy==1.24.4
onnxruntime==1.15.1
opencv-python-headless==4.8.0.74
packaging==23.1
Pillow==10.0.0
platformdirs==3.8.0
pooch==1.7.0
protobuf==4.23.3
pyinstaller==5.13.0
pyinstaller-hooks-contrib==2023.4
PyMatting==1.1.8
PyWavelets==1.4.1
rembg==2.0.49
requests==2.31.0
scikit-image==0.21.0
scipy==1.11.1
sympy==1.12
tifffile==2023.4.12
tqdm==4.65.0
urllib3==2.0.3
